Add video extraction test
