+++ src/vision/sdxl.py
@@ -46,11 +46,3 @@ def change_image(prompt, init_image=None, init_file=None, filename=None, gpu_id=
         return image
 
 
-def test_make_image():
-    prompt = "A cinematic shot of a baby racoon wearing an intricate italian priest robe."
-    make_image(prompt, filename="output_p2i.png")
-
-
-def test_change_image():
-    init_file = "https://huggingface.co/datasets/huggingface/documentation-images/resolve/main/diffusers/cat.png"
-    change_image(init_file=init_file, prompt="cat wizard, gandalf, lord of the rings, detailed, fantasy, cute, adorable, Pixar, Disney, 8k", filename="output_i2i.png")
